removed checking for funder names in slot names for GroupSlot + lint
