+++ chatsky/slots/slots.py
@@ -8,18 +8,24 @@ from __future__ import annotations
 
 import asyncio
 import re
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from typing import Callable, Any, Awaitable, TYPE_CHECKING, Union, Optional, Dict
-from typing_extensions import TypeAlias, Annotated
+from typing_extensions import Annotated
 import logging
-from functools import reduce
-from string import Formatter
 
 from pydantic import BaseModel, model_validator, Field, field_serializer, field_validator
 
 from chatsky.utils.devel.async_helpers import wrap_sync_function_in_async
 from chatsky.utils.devel.json_serialization import pickle_serializer, pickle_validator
-from chatsky.slots.base_slots import ExtractedSlot, BaseSlot, SlotNotExtracted, KwargOnlyFormatter
+from chatsky.slots.base_slots import (
+    ExtractedSlot,
+    BaseSlot,
+    SlotNotExtracted,
+    KwargOnlyFormatter,
+    SlotName,
+    recursive_getattr,
+    recursive_setattr,
+)
 
 if TYPE_CHECKING:
     from chatsky.core import Context, Message
@@ -161,43 +167,36 @@ class ValueSlot(BaseSlot, frozen=True):
         )
 
 
-class GroupSlot(BaseSlot, extra="allow", frozen=True):
+class GroupSlot(BaseSlot, frozen=True):
     """
     Base class for :py:class:`~.RootSlot` and :py:class:`~.GroupSlot`.
     """
 
     value_format: str = None
-    __pydantic_extra__: Dict[str, Annotated[Union["GroupSlot", "ValueSlot"], Field(union_mode="left_to_right")]]
+    slots: Dict[str, Annotated[Union["GroupSlot", "ValueSlot"], Field(union_mode="left_to_right")]]
 
     def __init__(self, **kwargs):  # supress unexpected argument warnings
         super().__init__(**kwargs)
 
     @model_validator(mode="after")
-    def __check_extra_field_names__(self):
+    def __check_slot_names__(self):
         """
-        Extra field names cannot be dunder names or contain dots.
+        Slot names cannot contain dots.
         """
-        for field in self.__pydantic_extra__.keys():
+        for field in self.slots.keys():
             if "." in field:
                 raise ValueError(f"Extra field name cannot contain dots: {field!r}")
-            if field.startswith("__") and field.endswith("__"):
-                raise ValueError(f"Extra field names cannot be dunder: {field!r}")
         return self
 
     async def get_value(self, ctx: Context) -> ExtractedGroupSlot:
-        child_values = await asyncio.gather(*(child.get_value(ctx) for child in self.__pydantic_extra__.values()))
+        child_values = await asyncio.gather(*(child.get_value(ctx) for child in self.slots.values()))
         return ExtractedGroupSlot(
             value_format=self.value_format,
-            **{
-                child_name: child_value
-                for child_value, child_name in zip(child_values, self.__pydantic_extra__.keys())
-            }
+            slots={child_name: child_value for child_value, child_name in zip(child_values, self.slots.keys())},
         )
 
     def init_value(self) -> ExtractedGroupSlot:
-        return ExtractedGroupSlot(
-            **{child_name: child.init_value() for child_name, child in self.__pydantic_extra__.items()}
-        )
+        return ExtractedGroupSlot(**{child_name: child.init_value() for child_name, child in self.slots.items()})
 
 
 class RegexpSlot(ValueSlot, frozen=True):
@@ -270,3 +269,123 @@ class FunctionSlot(ValueSlot, frozen=True):
     async def extract_value(self, ctx: Context) -> Union[Any, SlotNotExtracted]:
         return await wrap_sync_function_in_async(self.func, ctx.last_request)
 
+
+class SlotManager(BaseModel):
+    """
+    Provides API for managing slots.
+
+    An instance of this class can be accessed via ``ctx.framework_data.slot_manager``.
+    """
+
+    slot_storage: ExtractedGroupSlot = Field(default_factory=ExtractedGroupSlot)
+    """Slot storage. Stored inside ctx.framework_data."""
+    root_slot: GroupSlot = Field(default_factory=GroupSlot, exclude=True)
+    """Slot configuration passed during pipeline initialization."""
+
+    def set_root_slot(self, root_slot: GroupSlot):
+        """
+        Set root_slot configuration from pipeline.
+        Update extracted slots with the new configuration:
+
+        New slots are added with their :py:meth:`~.BaseSlot.init_value`.
+        Old extracted slot values are preserved only if their configuration did not change.
+        That is if they are still present in the config and if their fundamental type did not change
+        (i.e. `GroupSlot` did not turn into a `ValueSlot` or vice versa).
+
+        This method is called by pipeline and is not supposed to be used otherwise.
+        """
+        self.root_slot = root_slot
+        new_slot_storage = root_slot.init_value()
+        new_slot_storage.update(self.slot_storage)
+        self.slot_storage = new_slot_storage
+
+    def get_slot(self, slot_name: SlotName) -> BaseSlot:
+        """
+        Get slot configuration from the slot name.
+
+        :raises KeyError: If the slot with the specified name does not exist.
+        """
+        slot = recursive_getattr(self.root_slot, slot_name)
+        if isinstance(slot, BaseSlot):
+            return slot
+        raise KeyError(f"Could not find slot {slot_name!r}.")
+
+    async def extract_slot(self, slot_name: SlotName, ctx: Context, success_only: bool) -> None:
+        """
+        Extract slot `slot_name` and store extracted value in `slot_storage`.
+
+        :raises KeyError: If the slot with the specified name does not exist.
+
+        :param slot_name: Name of the slot to extract.
+        :param ctx: Context.
+        :param success_only: Whether to store the value only if it is successfully extracted.
+        """
+        slot = self.get_slot(slot_name)
+        value = await slot.get_value(ctx)
+
+        if value.__slot_extracted__ or success_only is False:
+            recursive_setattr(self.slot_storage, slot_name, value)
+
+    async def extract_all(self, ctx: Context):
+        """
+        Extract all slots from slot configuration `root_slot` and set `slot_storage` to the extracted value.
+        """
+        self.slot_storage = await self.root_slot.get_value(ctx)
+
+    def get_extracted_slot(self, slot_name: SlotName) -> Union[ExtractedValueSlot, ExtractedGroupSlot]:
+        """
+        Retrieve extracted value from `slot_storage`.
+
+        :raises KeyError: If the slot with the specified name does not exist.
+        """
+        slot = recursive_getattr(self.slot_storage, slot_name)
+        if isinstance(slot, ExtractedSlot):
+            return slot
+        raise KeyError(f"Could not find slot {slot_name!r}.")
+
+    def is_slot_extracted(self, slot_name: str) -> bool:
+        """
+        Return if the specified slot is extracted.
+
+        :raises KeyError: If the slot with the specified name does not exist.
+        """
+        return self.get_extracted_slot(slot_name).__slot_extracted__
+
+    def all_slots_extracted(self) -> bool:
+        """
+        Return if all slots are extracted.
+        """
+        return self.slot_storage.__slot_extracted__
+
+    def unset_slot(self, slot_name: SlotName) -> None:
+        """
+        Mark specified slot as not extracted and clear extracted value.
+
+        :raises KeyError: If the slot with the specified name does not exist.
+        """
+        self.get_extracted_slot(slot_name).__unset__()
+
+    def unset_all_slots(self) -> None:
+        """
+        Mark all slots as not extracted and clear all extracted values.
+        """
+        self.slot_storage.__unset__()
+
+    def fill_template(self, template: str) -> Optional[str]:
+        """
+        Fill `template` string with extracted slot values and return a formatted string
+        or None if an exception has occurred while trying to fill template.
+
+        `template` should be a format-string:
+
+        E.g. "Your username is {profile.username}".
+
+        For the example above, if ``profile.username`` slot has value "admin",
+        it would return the following text:
+        "Your username is admin".
+        """
+        try:
+            return KwargOnlyFormatter().format(template, **dict(self.slot_storage.__pydantic_extra__.items()))
+        except Exception as exc:
+            logger.exception("An exception occurred during template filling.", exc_info=exc)
+            return None
